requirements.txt
python-dotenv
discord-py-interactions==5.15.0
pydantic
colorlog
requests